Big Refactoring, Removing unnecessary code.
Now that we are able to have a channel for Dictionary and YouTube videos is not necessary to have a bot to strictly do these things via database.
It hurts me, because I spent a lot of time and research but... It's ok, at least I could learn and it's not totally lost since we have commits.